+++ libdevcore/Exceptions.cpp
@@ -24,10 +24,14 @@ using namespace dev;
 
 char const* Exception::what() const noexcept
 {
+	// Return the comment if available.
 	if (string const* cmt = comment())
-		return cmt->c_str();
-	else
-		return nullptr;
+		return cmt->data();
+	
+	// Fallback to base what().
+	// Boost accepts nullptr, but the C++ standard doesn't
+	// and crashes on some platforms.
+	return std::exception::what();
 }
 
 string Exception::lineInfo() const